Factor alloc_utils out.
Currently, we statically allocate resources--namely, tap interfaces--for
a given number of instances. This means that we are creating many
interfaces that we may not actually use.

The previous mechanism partially implemented to solve this problem was
to delegate allocation to a sidecar daemon, allocd, that would receive
RPCs to signal that resources should be allocated and tore down.
However, we can use a simpler design that creates interfaces before the
vmm starts and tears them down after the vmm exits.

The existing implementation shells out to use the platform ip(8),
ebtables(8), and dnsmasq(8) commands to make the necessary network
changes. While we could reimplemnt these using netlink, for now, it makes
sense to just expose the existing implementation from allocd so we can
borrow it later.