Incrementally lock files on creation.
Since we currently statically allocate tap interfaces for the maximum
number of instances we anticipate running, the current code tries to
acquire a number of instance locks for the number of tap interfaces
currently on the system.

If we want to only acquire tap interfaces when instances are created,
this scheme has to be modified. Instead, we only acquire as many locks
for the number of instances requested to be created.