cvdalloc binary.
This process--cvdalloc--will eventually be invoked from run_cvd, intended
to be invoked before the vmm starts. Because modifying the network
requires elevated permissions, we must be setuid root. Network resources
are allocated using the existing allocd implementation we exposed
earlier.

We must ensure that interfaces are set up and available before the vmm
starts. To synchronize around this, run_cvd will call socketpair(2)
twice for signaling important events (essentially, using these like
semaphores), the first event being that cvdalloc has finished
allocating resources and thus the vmm can properly start, and the other
event being that the vmm has exited and cvdalloc can tear down the
resources that it had originally created.

The run_cvd process manages the lifecycle of the instance and is
responsible for starting and stopping ancillary binaries--such as
cvdalloc--that make the guest perform as intended. As the instance is
being stopped, there are mechanisms that allow us to cleanly stop
cvdalloc so run_cvd could correctly signal that resources could be torn
down. However, since shutdown order is not necessarily guaranteed, it
can be difficult to ensure that cvdalloc's shutdown signal is received
after the vmm exits.

Instead, cvdalloc forks and then blocks on the teardown socketpair. Once
run_cvd--which holds the other end of the socketpair--exits, that end is
implictly shut down, which unblocks cvdalloc to complete the teardown.
This ensures teardown executes once run_cvd exits. This is not a
guarantee that teardown executes once the vmm exits; it may be the case
that run_cvd exits abnormally, triggering cvdalloc to attempt to
shutdown, but the vmm may still be running. For now, we ignore this
case, but later we can use retries to try and mitigate this problem.

This version is very basic which will be iterated on in future commits;
it is not wired up to anything yet.