Wire cvdalloc up to run_cvd.
Recall that we need run_cvd to invoke socketpair(2) to ensure that
we can signal cvdalloc that resources can be torn down, and that the
equivalent of shutdown(2) occurs on the run_cvd end of the socketpair
when that process exits.

We also have a socketpair to wait on cvdalloc completing its allocation
so that we know we can start up the vmm without issue. To do that, we
need to override the WaitForAvailability method provided through
VmmDependencyCommand; all classes implementing VmmDependencyCommand will
have their WaitForAvailability methods called synchronously before the
vmm is started, which achieves the intended result.

The instance id is also passed to cvdalloc here so that cvdalloc can
just simply trust that it is the appropriate instance ID and let that be
managed elsewhere.