Validate cvdalloc binary file permissions.
If we are trying to run cvdalloc with a freshly-built binary, then it
will not work since it will not be minted already setuid root for
obvious reasons. However, we should at least detect this early so that
you don't find out later when the vmm dies because cvdalloc couldn't
set up interfaces because it lacked permissions.