Validate cvdalloc binary file permissions.
If we are trying to run cvdalloc with a freshly-built binary, then it
will not work since it will not be minted already setuid root for
obvious reasons. However, we should at least detect this somewhat early
so that you don't find out later when the vmm dies because cvdalloc
couldn't set up interfaces because it lacked permissions.

We likely can't detect this earlier than here, because that requires
consuming flags that we want to eventually pass on.