Pass DefaultNetworkInterfaces the config.
The IfaceConfig as returned by DefaultNetworkInterfaces informs the vmm
about the network interfaces it can use, including the interface names,
which embed the instance number. Instead of passing the interface number
directly, pass the CuttlefishConfig instead.

We will soon use the flag we generated earlier to tell the vmm to expect
interfaces with different names when we are using cvdalloc, so that
cvdalloc doesn't try to modify existing tap interfaces generated by
the cuttlefish-host-resources service. We could that this by detecting if
the service is running, but doing that effectively will be a pain.