Name cvdalloc-managed interfaces differently.
Tell the vmm to expect cvdalloc-managed tap interfaces to be named
differently from preallocated tap interfaces. This way, we avoid
conflicts on a host where cuttlefish-host-resources may be running.

While we are doing this, we remove the padding requirements for the
interface number in the naming scheme for cvdalloc-managed interfaces.
This makes the interface naming more idiomatic and avoids the potential
problem in the future if we have more than 99 instances on a host.