Use incremental locking when use_cvdalloc.
The existing locking scheme examines the number of preexisting tap
interfaces on the system as created by cuttlefish-host-resources, then
acquires the number of locks as there are tap interfaces. This scheme
won't work when we want to incrementally create tap interfaces. Instead,
acquire as many instance locks as requested when we use_cvdalloc.

The case where instance numbers can be explicitly requested is slightly
interesting; here we use TryAcquireLocks. Arguably, this could be used
in the non-use_cvdalloc case, but in that case, the instance numbers are
bounded by the number of tap interfaces, so will have different results.
Branch the behavior all the same, though.