Use incremental locking when use_cvdalloc.
The existing locking scheme examines the number of preexisting tap
interfaces on the system as created by cuttlefish-host-resources, then
acquires the number of locks as there are tap interfaces. This scheme
won't work when we want to incrementally create tap interfaces. Instead,
acquire as many instance locks as requested when we use_cvdalloc.