cvdalloc supports non-bridged Wi-Fi.
Most of this is plumbing the right arguments to OpenWRT to specify a
network for the instance, as well as creating the non-bridged tap
interfaces.

The existing behavior is to use non-bridged Wi-Fi if the tap device
exists and the 802.11 simulator is enabled; since the tap device doesn't
exist by default for cvdalloc, we just use the use_cvdalloc flag as a
proxy for that condition.