Package the defaults file.
Loading new feature flag defaults from a file enables deployment of new
features by deploying a new defaults file rather than deploying the
entire software package.

One mechanism that we can effect this deployment is by using regular
Debian packaging; here, we introduce a new package `cuttlefish-defaults`
for rolling out feature flag default changes. When we want to
experimentally deploy a new feature associated with a feature flag, the
defaults file is changed and a new release of this package is cut. This
new version of the package can reside in an unstable repository at a
different cadence from the software package, which can go to stable
earlier.