In GCE, derive defaults from metadata.
Cloud infrastructure often users to deploy instances with associated
metadata, for purposes like associating data such as configuration, or
parameters describing instance purposes. This provides a natural means
for feature flag deployment.

We introduce a new defaults binary that runs in GCE instances, deployed
with the `cuttlefish-integration` package, and runs at boot. This
binary can be invoked with an optional flag that looks for a particular
key-value pair in the instance metadata, and if true, initializes the
defaults file with statically defined content. Otherwise, the binary
will load the defaults file with the content derived from metadata
verbatim.